requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
openai>=1.10.0
git+https://github.com/NickWaterton/samsung-tv-ws-api.git